Moving brie to pages/home/brie-slice, adding .DS_Store to .gitignore
