+++ .gitignore
@@ -16,3 +16,4 @@ node_modules
 npm-debug.log*
 pids
 .env
+.DS_Store
