--- site/components/brie/style.css
@@ -1,30 +1,0 @@
-@value red, white, black from "../../css/_colors.css";
-@value mediumScreen from "../../css/_breakpoints.css";
-
-.brieContainer {
-  padding: 30px 20px;
-  composes: dividerWhite from "../../css/_divider.css";
-  background-color: white;
-  text-align: center;
-}
-
-.heading {
-  composes: serifFontS from "../../css/typography/_fonts.css";
-}
-
-.brie {
-  margin: 30px 0px;
-  display: block;
-}
-
-.featureBtn {
-  line-height: 1;
-  composes: primaryButton from "../../css/_links.css";
-  display: block;
-}
-
-@media mediumScreen {
-  .featureBtn {
-    display: inline-block;
-  }
-}
